usb: Create combined usb-normalboot and usb-noadb
This file consists of usb-normalboot and usb-noadb combined.
The usb boot initialisation will be set by adb_setting.
For now to test the script it has been set to "1" (start adb)
Later this script will fetch the status of the adb setting from the
misc partition with the tool adbcheck, and set the variable adb_setting accordingly.