usb: Fixed wrong conditional setting
adbcheck returns 0 as result when ADB should be disabled.
In my initial test I had it set to 1 to disable ADB.
This patch fixes the wrong conditional test in the script.
So having the value 0 obtained from the misc partition make ADB disabled each boot.
Having it set to anything else makes ADB start at boot, so the problem can be triaged.