usb: Most minimalistic approach persistent ADB
This is the most minimalistic approach for switching ADB persistance.
Removed all the hsic helper files as they are not neccesary for this application.
Also "online" switching of adb state has been fixed in this release.
The "online" swtiching just like the ADBOFF command triggers a USB Reset for a short period of time.